Day 4
1.	First issue was the goFight function doesn’t work. It’s showing null which means no value.
Solution: All right. The problem was the “let”. WE already declared this as a global variable and declared this again which result to null or undefined. I just remove the let in the function.

2.	Second issue is that we’re going to round up the gold depends on the monster’s level.
Solution: gold+= Math.floor(monsters[fighting].level);
3.	The monsterStats.style.display doesn’t work. Either the synthax is wrong.
Solution: So basically none is declared with double quotes “”.
4.	Adding a condition where if we fight the dragon a function would be called.
Solution: The else if statement in the else if function was not necessary. Remove it.
5.	There is a bug on the health. Enemy doesn’t  die when health is below 0.
Solution: So the syntax of the condition is wrong. Place it in the monsterHealth condition function.